Moved extendTheme.js into source directory.
