+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import { ChakraProvider } from '@chakra-ui/react';
-import customTheme from '../exthendTheme';
+import customTheme from './extendTheme';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
